refactor and move cloudflare image transformation plugin
